+++ java/integration-tests/aws-sdk/agent-confmap/main.tf
@@ -7,7 +7,7 @@ locals {
 resource "aws_lambda_layer_version" "sdk_layer" {
   layer_name          = var.sdk_layer_name
   filename            = "${path.module}/../../../../opentelemetry-lambda/java/layer-javaagent/build/distributions/opentelemetry-javaagent-layer.zip"
-  compatible_runtimes = ["java8", "java8.al2", "java11"]
+  compatible_runtimes = ["java8.al2", "java11", "java17"]
   license_info        = "Apache-2.0"
   source_code_hash    = filebase64sha256("${path.module}/../../../../opentelemetry-lambda/java/layer-javaagent/build/distributions/opentelemetry-javaagent-layer.zip")
 }
@@ -16,7 +16,7 @@ resource "aws_lambda_layer_version" "collector_layer" {
   count               = var.enable_collector_layer ? 1 : 0
   layer_name          = var.collector_layer_name
   filename            = "${path.module}/../../../../opentelemetry-lambda/collector/build/opentelemetry-collector-layer-${local.architecture}.zip"
-  compatible_runtimes = ["nodejs14.x", "nodejs16.x", "nodejs18.x"]
+  compatible_runtimes =  ["nodejs16.x", "nodejs18.x", "nodejs20.x", "nodejs22.x"]
   license_info        = "Apache-2.0"
   source_code_hash    = filebase64sha256("${path.module}/../../../../opentelemetry-lambda/collector/build/opentelemetry-collector-layer-${local.architecture}.zip")
 }
@@ -25,6 +25,7 @@ module "hello-lambda-function" {
   source                     = "../../../../opentelemetry-lambda/java/sample-apps/aws-sdk/deploy/agent"
   name                       = var.function_name
   architecture               = var.architecture
+  runtime                    = var.runtime
   collector_layer_arn        = var.enable_collector_layer ? aws_lambda_layer_version.collector_layer[0].arn : null
   sdk_layer_arn              = aws_lambda_layer_version.sdk_layer.arn
   collector_config_layer_arn = aws_lambda_layer_version.collector_config_layer.arn
@@ -47,6 +48,7 @@ locals {
 auth:
   authenticator: sigv4auth
 endpoint: "${aws_prometheus_workspace.test_amp_workspace.prometheus_endpoint}api/v1/remote_write"
+add_metric_suffixes: false
 EOT
 }
 
@@ -80,7 +82,7 @@ receivers:
       http:
         endpoint: "localhost:4318"
 exporters:
-  logging:
+  debug:
   awsxray:
   prometheusremotewrite: $${${module.remote_configuration.configuration_uri}}
 
@@ -93,7 +95,7 @@ service:
       exporters: [awsxray]
     metrics:
       receivers: [otlp]
-      exporters: [logging, prometheusremotewrite]
+      exporters: [debug, prometheusremotewrite]
   telemetry:
     metrics:
       address: localhost:8888
@@ -123,6 +125,6 @@ resource "aws_lambda_layer_version" "collector_config_layer" {
   depends_on          = [data.archive_file.init]
   layer_name          = "custom-config-layer"
   filename            = "${path.module}/build/custom-config-layer.zip"
-  compatible_runtimes = ["java8", "java8.al2", "java11"]
+  compatible_runtimes = ["java8.al2", "java11", "java17"]
   license_info        = "Apache-2.0"
 }