+++ patch-upstream.sh
@@ -19,6 +19,7 @@ cp -rf adot/* opentelemetry-lambda/
 # Get current repo path
 CURRENT_DIR=$PWD
 
+
 # Move to the upstream OTel Lambda Collector folder where we will build a
 # collector used in each Lambda layer
 cd opentelemetry-lambda/collector
@@ -30,8 +31,7 @@ if [ -f $PATCH_OTEL_VERSION ]; then
     patch -p2 < $PATCH_OTEL_VERSION;
 fi
 
-# patch collector startup to remove HTTP and S3 confmap providers
-# and set ADOT-specific BuildInfo
+# patch collector startup to add httpsprovider
 patch -p2 < ../../collector.patch
 
 # patch manager.go to remove lambdacomponents attribute